Endre event-name og refaktorert Rivers som lagrer forespørsler
