+++ src/main/kotlin/no/nav/helsearbeidsgiver/bro/sykepenger/kafkatopic/spleis/Spleis.kt
@@ -28,7 +28,7 @@ object Spleis {
 
     @Serializable
     enum class Event {
-        TRENGER_OPPLYSNINGER_FRA_ARBEIDSGIVER,
+        TRENGER_OPPLYSNINGER_FRA_ARBEIDSGIVER_KOMPLETT,
         TRENGER_OPPLYSNINGER_FRA_ARBEIDSGIVER_BEGRENSET
     }
 }
